requirements.txt
langgraph==0.2.55
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
groq==0.13.0
easyocr==1.7.1
tavily-python==0.5.0
chromadb==0.5.20
sentence-transformers==3.3.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
streamlit==1.40.2
Pillow==11.0.0
httpx==0.28.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.3
